Create Stockloss Reporting
In order to build a stockloss dashbaord in Power BI, this SQL code creates 2 views which are then used to create the 2 tables which the dashboard is built from. The first half focuses on stock loss and the latter half focuses on incidents. I build views so that there can be a place where I can make amendments without affecting the data. The table holds the static data in order to improve performance