Create SQL - Fitting Room Metrics
I was tasked with making a report which shows the performance of each store with smart fitting rooms. Stakeholders wanted a report which showed the store performance for the most recent period, average performance in the last year and the difference. Doing this required using advanced logic and surrogate keys on weeks to create rolling 13 period windows. The metrics requested were:

1. Capture Rate: Proportion of customers using the fitting rooms
2. Fitting room visits --> sales: Proportion of fitting room visits which led to a sale
3. Fitting room requests --> sales: Of all the requests made, how many led to a sale? 
4. Missed Request % --> Of all requests, how many were missed by staff. 

This report will be used to identify stores where installing the smart fitting rooms were a success and ones where it is not profitable.