Add files via upload
This Power BI dashboard is built from 6 different tables which I created in Snowflake. It shows how stores with smart fitting rooms are performing in the most recent week, YTD and how they are performing in terms of missed requests. It also gives insight into the number of size up vs size down requests. This gives users a great deal of insight into customer behaviour and trends. The data is for the whole year of 2024.